requirements.txt
# ===base===
torch>=1.2.0
torchvision
numpy
opencv-python
pillow
matplotlib

# ===optional===
torchviz